fix(commandbar): split orphans helpers into their own file
CI lint blocked the previous push: commandbar_execute.go grew to 918
lines, past revive's 800-line file-length-limit. Move the
:orphans command and its alias/preset lookup tables to
commandbar_orphans.go — a coherent unit that stands on its own. No
behavior change; commandbar_execute.go drops to 794 lines.
